fix: linux rope_apply precision + browser shader-f16 fallback
Three CI failures uncovered now that the matrix actually executes
the conformance tests:

1) Vulkan-direct rope_apply.comp.glsl — fp16 drift > 1 ULP

   maxAbs=9.77e-4 maxUlp=44.00 >1ULP=27/512 vs the browser fixture.
   Cause: glslang→SPIR-V keeps `float16_t * float16_t` strictly in
   fp16 between the multiply and the subtract; Tint→SPIR-V (which
   produces the fixture) keeps intermediates in fp32 and narrows
   once at the end. The pattern `a*c - b*s` cancels when the rope
   angle is small, so two roundings vs one diverges noticeably.

   Fix: explicit f32 intermediates in the GLSL, single narrowing
   on the final store. Matches Tint's behaviour.

   Verified locally: structural change only, Mac build unaffected
   (no Vulkan path on Mac). CI on Linux runner is the actual proof.

2) Browser auto-fallback didn't probe shader-f16

   PrivacyFilter.create() with no `backend` option picked WebGPU
   whenever `navigator.gpu` existed — even on adapters without
   shader-f16. WebGpuBackend.warmup() then threw at requestDevice.
   This is a real user-facing bug, not just a CI workaround:
   software adapters, headless chromium, and some older Mesa
   drivers don't expose shader-f16, and they should silently
   fall through to WASM.

   Fix: probe `adapter.features.has("shader-f16")` in the auto
   path and route to WASM when missing. Explicit `backend: "webgpu"`
   skips the probe so warmup throws with the actual WebGPU error.

3) Eight specs explicitly need WebGPU + shader-f16

   webgpu-{coldstart,forward,matmul,profile,smoke}, bench, and
   the two conformance specs (bench-shaders, dump-fixtures) all
   either request `backend: "webgpu"` or talk to WebGPU directly.
   Added a tests/browser/helpers/skip-without-shader-f16.ts probe
   wired via test.beforeEach. On Mac dev (shader-f16 present)
   they run as before; on CI's software adapter they skip cleanly
   instead of erroring.

+++ packages/textsift/src/native/vulkan/shaders/rope_apply.comp.glsl
@@ -40,11 +40,18 @@ void main() {
     float16_t a = qk[head_base + 2u * p];
     float16_t b = qk[head_base + 2u * p + 1u];
 
-    float16_t ac = a * c;
-    float16_t bs = b * s;
-    float16_t bc = b * c;
-    float16_t as_ = a * s;
+    // Compute the rotation in f32 and narrow once at the end. The
+    // browser fixture (Tint→SPIR-V) keeps intermediates in f32 across
+    // multiply→subtract, which matters here: `a*c - b*s` cancels when
+    // the rotation angle is small, and rounding twice at f16 (once after
+    // each multiply, once after the subtract) drifts up to ~44 ULPs vs
+    // the fixture. glslang→SPIR-V doesn't promote on its own, so we do
+    // it explicitly. Single narrowing rounding matches Tint's behaviour.
+    float ac = float(a) * float(c);
+    float bs = float(b) * float(s);
+    float bc = float(b) * float(c);
+    float as_ = float(a) * float(s);
 
-    qk[head_base + 2u * p]      = ac - bs;
-    qk[head_base + 2u * p + 1u] = bc + as_;
+    qk[head_base + 2u * p]      = float16_t(ac - bs);
+    qk[head_base + 2u * p + 1u] = float16_t(bc + as_);
 }
